+++ app/api/chemotion/reaction_api.rb
@@ -55,7 +55,6 @@ module ReactionHelpers
       purification_solvent: Array(material_attributes['purification_solvents']).map { |m| OSample.new(m) },
       product: Array(material_attributes['products']).map { |m| OSample.new(m) }
     }
-
     ActiveRecord::Base.transaction do
       included_sample_ids = []
       materials.each do |material_group, samples|
@@ -66,7 +65,7 @@ module ReactionHelpers
           sample.reference = false if material_group === 'solvent' && sample.reference == true
           # create new subsample
           if sample.is_new
-            if sample.is_split && sample.parent_id
+            if sample.parent_id && material_group != 'products'
               parent_sample = Sample.find(sample.parent_id)
 
               # TODO: extract subsample method
@@ -443,7 +442,7 @@ module Chemotion
         attributes.delete(:container)
         attributes.delete(:segments)
 
-        collection = Collection.find(collection_id)
+        collection = current_user.collections.where(id: collection_id).take
         attributes[:created_by] = current_user.id
         reaction = Reaction.create!(attributes)
         recent_ols_term_update('rxno', [params[:rxno]]) if params[:rxno].present?
@@ -477,8 +476,19 @@ module Chemotion
         reaction.container = update_datamodel(container_info)
         reaction.save!
         reaction.save_segments(segments: params[:segments], current_user_id: current_user.id)
-        CollectionsReaction.create(reaction: reaction, collection: collection)
-        CollectionsReaction.create(reaction: reaction, collection: Collection.get_all_collection_for_user(current_user.id))
+        CollectionsReaction.create(reaction: reaction, collection: collection) if collection.present?
+
+        is_shared_collection = false
+        unless collection.present?
+          sync_collection = current_user.all_sync_in_collections_users.where(id: collection_id).take
+          if sync_collection.present?
+            is_shared_collection = true
+            CollectionsReaction.create(reaction: reaction, collection: Collection.find(sync_collection['collection_id']))
+            CollectionsReaction.create(reaction: reaction, collection: Collection.get_all_collection_for_user(sync_collection['shared_by_id']))
+          end
+        end
+
+        CollectionsReaction.create(reaction: reaction, collection: Collection.get_all_collection_for_user(current_user.id)) unless is_shared_collection
         CollectionsReaction.update_tag_by_element_ids(reaction.id)
         if reaction
           if attributes['origin'] && attributes['origin']['short_label']