+++ app/packs/src/libHome/NavNewSession.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Button, FormGroup, FormControl, Glyphicon, Navbar, NavItem, Nav, OverlayTrigger, Tooltip } from 'react-bootstrap';
 
-const NewSession = ({ authenticityToken }) => (
+const NewSession = ({ authenticityToken, omniauthProviders }) => (
   <div>
     <Nav pullRight><NavItem href="/users/sign_up"> or Sign Up </NavItem></Nav>
     <Navbar.Form pullRight>
@@ -22,11 +22,32 @@ const NewSession = ({ authenticityToken }) => (
         </Button>
       </form>
     </Navbar.Form>
+    {
+      omniauthProviders && omniauthProviders.includes('orcid') && <Nav pullRight>
+        <NavItem href="/users/auth/orcid" className="omniauth-navitem">
+          Login ORCID
+        </NavItem>
+      </Nav>
+    }
+    {
+      omniauthProviders && omniauthProviders.includes('github') && <Nav pullRight>
+        <NavItem href="/users/auth/github" className="omniauth-navitem">
+          Login GitHub
+        </NavItem>
+      </Nav>
+    }
+    {
+      omniauthProviders && omniauthProviders.includes('openid_connect') && <Nav pullRight>
+        <NavItem href="/users/auth/openid_connect" className="omniauth-navitem">
+          &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Login OIDC
+        </NavItem>
+      </Nav>
+    }
   </div>
 );
 
 NewSession.propTypes = {
-  authenticityToken: PropTypes.string.isRequired,
+  authenticityToken: PropTypes.string.isRequired
 };
 
 export default NewSession;